# swagger_openapi_saludvital.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión Médica - Salud Vital API
  description: |
    API REST para el sistema de gestión médica integral. 
    Permite la gestión de pacientes, médicos, citas médicas y recetas.
    
    ## Usuarios Predefinidos para Pruebas:
    
    ### 👑 **ADMIN**
    - **Email**: `admin@saludvital.com`
    - **Password**: `admin123`
    
    ### 🩺 **MÉDICO** 
    - **Email**: `medico@saludvital.com`
    - **Password**: `medico123`
    
    ### 👤 **PACIENTE**
    - **Email**: `paciente@saludvital.com`
    - **Password**: `paciente123`
    
    ## Flujo de Autenticación:
    1. Hacer login con alguno de los usuarios predefinidos
    2. Copiar el `accessToken` de la respuesta
    3. Usar en header: `Authorization: Bearer {accessToken}`
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@saludvital.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api
    description: Servidor de Desarrollo
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token. Ejemplo 'Bearer {token}'
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "paciente@saludvital.com"
        password:
          type: string
          example: "paciente123"
    SignUpRequest:
      type: object
      required:
        - name
        - email
        - password
        - numeroIdentificacion
        - fechaNacimiento
      properties:
        name:
          type: string
          example: "Juan Pérez"
        email:
          type: string
          format: email
          example: "juan.perez@email.com"
        password:
          type: string
          minLength: 6
          example: "123456"
        numeroIdentificacion:
          type: string
          example: "12345678"
        fechaNacimiento:
          type: string
          format: date
          example: "1990-05-15"
    JwtAuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9..."
        tokenType:
          type: string
          example: "Bearer"
        userInfo:
          $ref: '#/components/schemas/UserInfoResponse'
    UserInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Juan Pérez"
        email:
          type: string
          example: "juan.perez@email.com"
        roles:
          type: array
          items:
            type: string
          example: ["ROLE_PACIENTE"]
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operación exitosa"
        data:
          type: object
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error en la operación"
paths:
  /auth/login:
    post:
      tags:
        - 🔐 Autenticación
      summary: Iniciar sesión
      description: |
        Autentica usuario y obtiene token JWT.
        
        **Usuarios predefinidos:**
        - Admin: `admin@saludvital.com` / `admin123`
        - Médico: `medico@saludvital.com` / `medico123`
        - Paciente: `paciente@saludvital.com` / `paciente123`
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Login como Admin
                value:
                  email: "admin@saludvital.com"
                  password: "admin123"
              medico:
                summary: Login como Médico
                value:
                  email: "medico@saludvital.com"
                  password: "medico123"
              paciente:
                summary: Login como Paciente
                value:
                  email: "paciente@saludvital.com"
                  password: "paciente123"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - 🔐 Autenticación
      summary: Registrar paciente
      description: Registra nuevo paciente en el sistema
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '200':
          description: Registro exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Error en los datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/pacientes:
    get:
      tags:
        - ⚙️ Administración
      summary: Listar todos los pacientes
      description: Obtiene la lista completa de pacientes (solo admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Pacientes obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo admin)

    post:
      tags:
        - ⚙️ Administración
      summary: Registrar nuevo paciente
      description: Crea un nuevo paciente en el sistema (solo admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
                - email
                - password
                - numeroIdentificacion
                - fechaNacimiento
              properties:
                nombre:
                  type: string
                  example: "María García"
                email:
                  type: string
                  format: email
                  example: "maria.garcia@email.com"
                password:
                  type: string
                  example: "123456"
                numeroIdentificacion:
                  type: string
                  example: "87654321"
                fechaNacimiento:
                  type: string
                  format: date
                  example: "1985-10-20"
      responses:
        '200':
          description: Paciente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo admin)
  /admin/medicos:
    post:
      tags:
        - ⚙️ Administración
      summary: Registrar nuevo médico
      description: Crea un nuevo médico en el sistema (solo admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
                - apellido
                - numeroLicencia
                - email
                - password
                - especialidad
                - tarifaConsulta
              properties:
                nombre:
                  type: string
                  example: "Dr. Carlos"
                apellido:
                  type: string
                  example: "Rodríguez"
                numeroLicencia:
                  type: string
                  example: "MED123456"
                telefono:
                  type: string
                  example: "+1234567890"
                email:
                  type: string
                  format: email
                  example: "dr.carlos@hospital.com"
                password:
                  type: string
                  example: "123456"
                especialidad:
                  type: string
                  enum: [CARDIOLOGIA, DERMATOLOGIA, GASTROENTEROLOGIA, GINECOLOGIA, NEUROLOGIA, OFTALMOLOGIA, PEDIATRIA, PSIQUIATRIA, TRAUMATOLOGIA, UROLOGIA, MEDICINA_GENERAL]
                  example: "CARDIOLOGIA"
                tarifaConsulta:
                  type: number
                  format: decimal
                  example: 150.00
      responses:
        '200':
          description: Médico creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo admin)
...